+++ src/App.jsx
@@ -1,7 +1,6 @@
 import React, { Component } from 'react';
 
-import Scene from './scene.js'
-// import "./App.css";
+import Scene from './Scene.jsx'
 
 class App extends Component {
   render() {
